Changed Get-CohesityProtectionSource cmdlet to only return top-level (rootNodes) sources.
- Only return top-level protection sources to keep it simple
- Can get a specific protection source based on id
- Can filter based on environments
- Will have a separate cmdlet to get other objects under the protection sources.
